ObjectLayer. Updated objectlayerservice, add unit tests for objectlayer

+++ Core/Options/AppSettings.cs
@@ -2,7 +2,7 @@
 {
     public class AppSettings
     {
-        public double MapWidth { get; set; } = 1000;
-        public double MapHeight { get; set; } = 1000;
+        public int MapWidth { get; set; } = 1000;
+        public int MapHeight { get; set; } = 1000;
     }
 }
